+++ src/main/java/org/recast/Recast/Include/rcEdge.java
@@ -1,7 +1,7 @@
 package org.recast.Recast.Include;
 
 public class rcEdge {
-    public short vert[] = new short[2];
+    public int vert[] = new int[2];
     public short polyEdge[] = new short[2];
     public short poly[] = new short[2];
 }
