+++ src/main/java/org/recast/Recast/Include/rcSweepSpan.java
@@ -1,8 +1,8 @@
 package org.recast.Recast.Include;
 
 public class rcSweepSpan {
-    public short rid;	// row id
-    public short id;	// region id
+    public int rid;	// row id
+    public int id;	// region id
     public short ns;	// number samples
-    public short nei;	// neighbour id
+    public int nei;	// neighbour id
 }
